ajout du middleware

+++ controllers/bars.controller.js
@@ -1,3 +1,4 @@
+const { validationResult } = require("express-validator");
 const Bars = require("../models/bars.model");
 
 const BarsController = {
@@ -9,7 +10,7 @@ const BarsController = {
    */
   all: async (req, res) => {
     try {
-      console.log("bar_list");
+      // fetch all bars
       const bar_list = await Bars.findAll({});
       if (!bar_list) {
         return res.status(404).json({
@@ -23,8 +24,7 @@ const BarsController = {
       });
     } catch (error) {
       return res.status(500).json({
-        message: "Request failed",
-        error: error.message,
+        message: `Error: ${error}}`,
       });
     }
   },
@@ -35,18 +35,30 @@ const BarsController = {
    * - data?: array | object
    */
   one: async (req, res) => {
-    const { id_bar } = req.params;
-    const bar = await Bars.findByPk(id_bar);
-    if (!bar) {
-      return res.status(404).json({
-        message: "Error: Bar not found",
-      });
+    // Verify if middleware catch some errors
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+      return res.status(400).json({ errors: errors.array() });
     }
 
-    return res.status(200).json({
-      message: "Request successfully completed",
-      data: bar,
-    });
+    try {
+      const { id_bar } = req.params;
+      const bar = await Bars.findByPk(id_bar);
+      if (!bar) {
+        return res.status(404).json({
+          message: "Error: Bar not found",
+        });
+      }
+
+      return res.status(200).json({
+        message: "Request successfully completed",
+        data: bar,
+      });
+    } catch (error) {
+      return res.status(500).json({
+        message: `Error: ${error}}`,
+      });
+    }
   },
   /**
    * Créer un nouveau bar
@@ -55,42 +67,54 @@ const BarsController = {
    * - data?: array | object
    */
   create: async (req, res) => {
-    const { name, adresse, tel, email, description } = req.body;
-    // Verify the existence of the bar
-    const bar_occ = await Bars.findOne({
-      where: {
+    // Verify if middleware catch some errors
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+      return res.status(400).json({ errors: errors.array() });
+    }
+
+    try {
+      const { name, adresse, tel, email, description } = req.body;
+      // Verify the existence of the bar
+      const bar_occ = await Bars.findOne({
+        where: {
+          name: name,
+          adresse: adresse,
+          tel: tel,
+          email: email,
+        },
+      });
+      if (bar_occ) {
+        return res.status(409).json({
+          message: "Error: Bar already exists",
+        });
+      }
+
+      // Create the bar
+      await Bars.create({
         name: name,
         adresse: adresse,
         tel: tel,
         email: email,
-      },
-    });
-    if (bar_occ) {
-      return res.status(409).json({
-        message: "Error: Bar already exists",
+        description: description,
+      })
+        .then(async (bar) => {
+          return res.status(201).json({
+            message: "Success: Request successfully completed",
+            data: bar,
+          });
+        })
+        .catch((error) => {
+          console.log(error);
+          return res.status(400).json({
+            message: "Error: Bars not created",
+          });
+        });
+    } catch (error) {
+      return res.status(500).json({
+        message: `Error: ${error}}`,
       });
     }
-
-    // Create the bar
-    await Bars.create({
-      name: name,
-      adresse: adresse,
-      tel: tel,
-      email: email,
-      description: description,
-    })
-      .then(async (bar) => {
-        return res.status(201).json({
-          message: "Success: Request successfully completed",
-          data: bar,
-        });
-      })
-      .catch((error) => {
-        console.log(error);
-        return res.status(400).json({
-          message: "Error: Bars not created",
-        });
-      });
   },
   /**
    * Mettre à jour un bar
@@ -99,38 +123,50 @@ const BarsController = {
    * - data?: array | object
    */
   update: async (req, res) => {
-    const { id_bar } = req.params;
-    const { name, adresse, tel, email, description } = req.body;
-
-    // Verify the existence of the bar
-    const bar_occ = await Bars.findByPk(id_bar);
-    if (!bar_occ) {
-      return res.status(404).json({
-        message: "Error: Bar not found",
-      });
+    // Verify if middleware catch some errors
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+      return res.status(400).json({ errors: errors.array() });
     }
 
-    // Update the bar
-    await Bars.update(
-      {
-        name: name,
-        adresse: adresse,
-        tel: tel,
-        email: email,
-        description: description,
-      },
-      { where: { id: id_bar } }
-    )
-      .then(() => {
-        return res.status(200).json({
-          message: "Success: Request successfully completed",
+    try {
+      const { id_bar } = req.params;
+      const { name, adresse, tel, email, description } = req.body;
+
+      // Verify the existence of the bar
+      const bar_occ = await Bars.findByPk(id_bar);
+      if (!bar_occ) {
+        return res.status(404).json({
+          message: "Error: Bar not found",
         });
-      })
-      .catch(() => {
-        return res.status(400).json({
-          message: "Error: Bars not updated",
+      }
+
+      // Update the bar
+      await Bars.update(
+        {
+          name: name,
+          adresse: adresse,
+          tel: tel,
+          email: email,
+          description: description,
+        },
+        { where: { id: id_bar } }
+      )
+        .then(() => {
+          return res.status(200).json({
+            message: "Success: Request successfully completed",
+          });
+        })
+        .catch(() => {
+          return res.status(400).json({
+            message: "Error: Bars not updated",
+          });
         });
+    } catch (error) {
+      return res.status(500).json({
+        message: `Error: ${error}}`,
       });
+    }
   },
   /**
    * Supprimer un bar
@@ -139,26 +175,38 @@ const BarsController = {
    * - data?: array | object
    */
   delete: async (req, res) => {
-    const { id_bar } = req.params;
+    // Verify if middleware catch some errors
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+      return res.status(400).json({ errors: errors.array() });
+    }
 
-    // Verify the existence of the bar
-    const bar_occ = await Bars.findByPk(id_bar);
-    if (!bar_occ) {
-      return res.status(404).json({
-        message: "Error: Bar not found",
+    try {
+      const { id_bar } = req.params;
+
+      // Verify the existence of the bar
+      const bar_occ = await Bars.findByPk(id_bar);
+      if (!bar_occ) {
+        return res.status(404).json({
+          message: "Error: Bar not found",
+        });
+      }
+      await Bars.destroy({ where: { id: id_bar } })
+        .then(() => {
+          return res.status(200).json({
+            message: "Success: Request successfully completed",
+          });
+        })
+        .catch(() => {
+          return res.status(400).json({
+            message: "Error: Bars not deleted",
+          });
+        });
+    } catch (error) {
+      return res.status(500).json({
+        message: `Error: ${error}}`,
       });
     }
-    await Bars.destroy({ where: { id: id_bar } })
-      .then(() => {
-        return res.status(200).json({
-          message: "Success: Request successfully completed",
-        });
-      })
-      .catch(() => {
-        return res.status(400).json({
-          message: "Error: Bars not deleted",
-        });
-      });
   },
 };
